Re-org mapping resources
