+++ lib/engine.js
@@ -17,7 +17,7 @@
 const _ = require('lodash');
 const micromatch = require('micromatch');
 const ast = require('./parser/ast');
-const { StructureMap } = require('./mapping-resources');
+const StructureMap = require('./mapping-resources/structure-map');
 
 class Engine {
   constructor (maps) {
