+++ lib/utils.js
@@ -24,7 +24,13 @@ function safeGetText (ctx, method) {
     ctx[method]().getText();
 }
 
+function stripDelimited (delimited) {
+  const string = delimited.toString();
+  return string.substring(1, string.length - 1);
+}
+
 module.exports = {
   typesByDefinitionAndPath,
-  safeGetText
+  safeGetText,
+  stripDelimited
 };
